requirements.txt
Django==1.7
django-suit==0.2.11
django_compressor==1.4
django-crispy-forms==1.4
django-bootstrap3-datetimepicker==2.2.3
django-tinymce==1.5.3
Pillow==2.5.3
Fabric==1.8.2
flake8==2.1
